register done

+++ routes/web.php
@@ -27,16 +27,17 @@ DELETE	    /photos/{photo}	        destroy	photos.destroy
 
 */
 Route::get('/', [ListingController::class, 'index'])->name('listings.index');
-Route::get('/listings/create', [ListingController::class, 'create'])->name('listings.create');
+Route::get('/listings/create', [ListingController::class, 'create'])->name('listings.create')->middleware('auth');
+Route::post('/listings/store', [ListingController::class, 'store'])->name('listings.store')->middleware('auth');;
 Route::get('/listings/{id}', [ListingController::class, 'show'])->name('listings.show');
-Route::post('/listings/store', [ListingController::class, 'store'])->name('listings.store');
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->name('listings.edit');
-Route::put('/listings/{listing}/update', [ListingController::class, 'update'])->name('listings.update');
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->name('listings.destroy');
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->name('listings.edit')->middleware('auth');;
+Route::put('/listings/{listing}/update', [ListingController::class, 'update'])->name('listings.update')->middleware('auth');;
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->name('listings.destroy')->middleware('auth');;
+// Route::get('/listings/manage', [ListingController::class, 'manage'])->name('listings.manage')->middleware('auth');;
 // Route::get('/', [ListingController::class, '']);
 
 Route::get('/register', [UserController::class, 'create']);
 Route::post('/user/store', [UserController::class, 'store']);
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login');
 Route::post('/authentication', [UserController::class, 'authentication']);
-Route::get('/logout', [UserController::class, 'logout']);
+Route::get('/logout', [UserController::class, 'logout'])->middleware('auth');;
